Move export declaration methods to common mixin.
